# vulkan: Optimize soft_max