Large soft_max could already saturate memory, but small/medium sizes were pretty slow. The bulk of the gains for them comes from using a smaller workgroup size, and making the workgroup size match the subgroup size also makes the barriers much cheaper.

Cache some values in locals to avoid refetching/recomputing. And stamp out a few "template instantiations" so smaller cases will fully unroll.

Add a missing early return for OOB rows. This happens when there are more than 512 rows and the dispatch is 512 x H.

These sizes I benchmarked came from a stable diffusion network I was looking at a while back.

```
Before:
  SOFT_MAX(type=f32,ne=[4096,4096,5,1],mask=0,scale=1.000000,max_bias=0.000000):                 728 runs -  1472.19 us/run -   655360 kB/run -  428.62 GB/s
  SOFT_MAX(type=f32,ne=[77,4096,5,1],mask=0,scale=1.000000,max_bias=0.000000):                  5448 runs -   351.30 us/run -    12320 kB/run -   33.45 GB/s
  SOFT_MAX(type=f32,ne=[1024,1024,10,1],mask=0,scale=1.000000,max_bias=0.000000):               4510 runs -   238.39 us/run -    81920 kB/run -  328.11 GB/s
  SOFT_MAX(type=f32,ne=[77,1024,10,1],mask=0,scale=1.000000,max_bias=0.000000):                10896 runs -   177.49 us/run -     6160 kB/run -   33.10 GB/s
  SOFT_MAX(type=f32,ne=[256,256,20,1],mask=0,scale=1.000000,max_bias=0.000000):                13108 runs -    92.21 us/run -    10240 kB/run -  105.92 GB/s
  SOFT_MAX(type=f32,ne=[64,64,20,1],mask=0,scale=1.000000,max_bias=0.000000):                  40955 runs -    29.06 us/run -      640 kB/run -   21.00 GB/s
  SOFT_MAX(type=f32,ne=[77,64,20,1],mask=0,scale=1.000000,max_bias=0.000000):                  40955 runs -    28.83 us/run -      770 kB/run -   25.47 GB/s
  SOFT_MAX(type=f32,ne=[4096,4096,5,1],mask=1,scale=1.000000,max_bias=0.000000):                 470 runs -  2164.47 us/run -   720896 kB/run -  320.70 GB/s
  SOFT_MAX(type=f32,ne=[77,4096,5,1],mask=1,scale=1.000000,max_bias=0.000000):                  4952 runs -   349.47 us/run -    13552 kB/run -   36.99 GB/s
  SOFT_MAX(type=f32,ne=[1024,1024,10,1],mask=1,scale=1.000000,max_bias=0.000000):               4301 runs -   247.73 us/run -    86016 kB/run -  331.54 GB/s
  SOFT_MAX(type=f32,ne=[77,1024,10,1],mask=1,scale=1.000000,max_bias=0.000000):                10376 runs -   176.68 us/run -     6468 kB/run -   34.92 GB/s
  SOFT_MAX(type=f32,ne=[256,256,20,1],mask=1,scale=1.000000,max_bias=0.000000):                12788 runs -    92.78 us/run -    10496 kB/run -  107.91 GB/s
  SOFT_MAX(type=f32,ne=[64,64,20,1],mask=1,scale=1.000000,max_bias=0.000000):                  40955 runs -    28.73 us/run -      656 kB/run -   21.78 GB/s
  SOFT_MAX(type=f32,ne=[77,64,20,1],mask=1,scale=1.000000,max_bias=0.000000):                  40955 runs -    28.92 us/run -      789 kB/run -   26.02 GB/s
  
After:
  SOFT_MAX(type=f32,ne=[4096,4096,5,1],mask=0,scale=1.000000,max_bias=0.000000):                 676 runs -  1558.61 us/run -   655360 kB/run -  404.85 GB/s
  SOFT_MAX(type=f32,ne=[77,4096,5,1],mask=0,scale=1.000000,max_bias=0.000000):                 46308 runs -    21.79 us/run -    12320 kB/run -  539.30 GB/s
  SOFT_MAX(type=f32,ne=[1024,1024,10,1],mask=0,scale=1.000000,max_bias=0.000000):               5330 runs -   198.13 us/run -    81920 kB/run -  394.80 GB/s
  SOFT_MAX(type=f32,ne=[77,1024,10,1],mask=0,scale=1.000000,max_bias=0.000000):                81720 runs -    12.47 us/run -     6160 kB/run -  471.09 GB/s
  SOFT_MAX(type=f32,ne=[256,256,20,1],mask=0,scale=1.000000,max_bias=0.000000):               108141 runs -     9.40 us/run -    10240 kB/run - 1038.94 GB/s
  SOFT_MAX(type=f32,ne=[64,64,20,1],mask=0,scale=1.000000,max_bias=0.000000):                 270303 runs -     3.71 us/run -      640 kB/run -  164.51 GB/s
  SOFT_MAX(type=f32,ne=[77,64,20,1],mask=0,scale=1.000000,max_bias=0.000000):                 278494 runs -     3.68 us/run -      770 kB/run -  199.55 GB/s
  SOFT_MAX(type=f32,ne=[4096,4096,5,1],mask=1,scale=1.000000,max_bias=0.000000):                 376 runs -  2963.52 us/run -   720896 kB/run -  234.23 GB/s
  SOFT_MAX(type=f32,ne=[77,4096,5,1],mask=1,scale=1.000000,max_bias=0.000000):                 49520 runs -    21.00 us/run -    13552 kB/run -  615.68 GB/s
  SOFT_MAX(type=f32,ne=[1024,1024,10,1],mask=1,scale=1.000000,max_bias=0.000000):               5083 runs -   204.13 us/run -    86016 kB/run -  402.34 GB/s
  SOFT_MAX(type=f32,ne=[77,1024,10,1],mask=1,scale=1.000000,max_bias=0.000000):                88196 runs -    12.03 us/run -     6468 kB/run -  513.00 GB/s
  SOFT_MAX(type=f32,ne=[256,256,20,1],mask=1,scale=1.000000,max_bias=0.000000):               102304 runs -    10.00 us/run -    10496 kB/run - 1001.46 GB/s
  SOFT_MAX(type=f32,ne=[64,64,20,1],mask=1,scale=1.000000,max_bias=0.000000):                 278494 runs -     3.61 us/run -      656 kB/run -  173.13 GB/s
  SOFT_MAX(type=f32,ne=[77,64,20,1],mask=1,scale=1.000000,max_bias=0.000000):                 270303 runs -     3.75 us/run -      789 kB/run -  200.55 GB/s
  
CUDA:
  SOFT_MAX(type=f32,ne=[4096,4096,5,1],mask=0,scale=1.000000,max_bias=0.000000):                 676 runs -  1491.82 us/run -   655360 kB/run -  422.98 GB/s
  SOFT_MAX(type=f32,ne=[77,4096,5,1],mask=0,scale=1.000000,max_bias=0.000000):                 21792 runs -    46.83 us/run -    12320 kB/run -  250.96 GB/s
  SOFT_MAX(type=f32,ne=[1024,1024,10,1],mask=0,scale=1.000000,max_bias=0.000000):               3280 runs -   333.83 us/run -    81920 kB/run -  234.31 GB/s
  SOFT_MAX(type=f32,ne=[77,1024,10,1],mask=0,scale=1.000000,max_bias=0.000000):                43584 runs -    24.65 us/run -     6160 kB/run -  238.38 GB/s
  SOFT_MAX(type=f32,ne=[256,256,20,1],mask=0,scale=1.000000,max_bias=0.000000):                42601 runs -    25.40 us/run -    10240 kB/run -  384.60 GB/s
  SOFT_MAX(type=f32,ne=[64,64,20,1],mask=0,scale=1.000000,max_bias=0.000000):                 311258 runs -     3.27 us/run -      640 kB/run -  186.94 GB/s
  SOFT_MAX(type=f32,ne=[77,64,20,1],mask=0,scale=1.000000,max_bias=0.000000):                 229348 runs -     4.44 us/run -      770 kB/run -  165.44 GB/s
  SOFT_MAX(type=f32,ne=[4096,4096,5,1],mask=1,scale=1.000000,max_bias=0.000000):                 470 runs -  2208.65 us/run -   720896 kB/run -  314.29 GB/s
  SOFT_MAX(type=f32,ne=[77,4096,5,1],mask=1,scale=1.000000,max_bias=0.000000):                 22284 runs -    48.65 us/run -    13552 kB/run -  265.73 GB/s
  SOFT_MAX(type=f32,ne=[1024,1024,10,1],mask=1,scale=1.000000,max_bias=0.000000):               3128 runs -   338.29 us/run -    86016 kB/run -  242.78 GB/s
  SOFT_MAX(type=f32,ne=[77,1024,10,1],mask=1,scale=1.000000,max_bias=0.000000):                41504 runs -    25.95 us/run -     6468 kB/run -  237.73 GB/s
  SOFT_MAX(type=f32,ne=[256,256,20,1],mask=1,scale=1.000000,max_bias=0.000000):                41561 runs -    25.66 us/run -    10496 kB/run -  390.13 GB/s
  SOFT_MAX(type=f32,ne=[64,64,20,1],mask=1,scale=1.000000,max_bias=0.000000):                 311258 runs -     3.25 us/run -      656 kB/run -  192.73 GB/s
  SOFT_MAX(type=f32,ne=[77,64,20,1],mask=1,scale=1.000000,max_bias=0.000000):                 221157 runs -     4.66 us/run -      789 kB/run -  161.51 GB/s
```

- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [x] Low
  - [ ] Medium
  - [ ] High