@0cc4m 2024-11-17
I can confirm this significantly improves softmax performance for small and medium sizes, but I also see a regression for large sizes. On your GPU and on my AMD Radeon Pro VII it seems to be minimal, but in the 4096, 4096, 5, 1 tests for Intel A770 and Nvidia RTX 3090 I see a big difference. Do you have an idea what causes this?

![AMD_Radeon_(TM)_Pro_VII_(RADV_VEGA20)_performance_comparison](https://github.com/user-attachments/assets/aee8eed4-5887-4237-957d-f8b5aedde3b0)
![Intel(R)_Arc(tm)_A770_Graphics_(DG2)_performance_comparison](https://github.com/user-attachments/assets/811d0820-a8c7-43e3-aab5-b34a6de6621a)
![NVIDIA_GeForce_RTX_3090_performance_comparison](https://github.com/user-attachments/assets/428c1790-0f85-49c2-8717-009461af2001)