@jeffbolznv 2024-11-17
The big difference between Ampere and Ada that comes to mind is the much larger L2 cache size on Ada. I'm guessing that with the larger workgroup size there are fewer rows being processed at a time and it can hit in the smaller L2 cache. I probably need to bring back the larger block size and use it when the rows are large enough. I have an RTX 3070 I can try this out on tomorrow.